sdk/mem0_bridge/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mem0-sgraal"
version = "0.1.0"
description = "Drop-in Sgraal preflight guard for Mem0 Memory"
requires-python = ">=3.9"
dependencies = ["mem0ai>=0.1.0", "requests>=2.28.0"]
license = "MIT"

[project.urls]
Homepage = "https://sgraal.com"
